# apps/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.DeleteLogsRequest:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_vo.LoginLogItem:
    properties:
      browser:
        type: string
      id:
        type: integer
      ipaddr:
        type: string
      location:
        type: string
      loginTime:
        type: string
      msg:
        type: string
      os:
        type: string
      status:
        type: integer
      username:
        type: string
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_vo.OperLogDetail:
    properties:
      businessType:
        type: integer
      costTime:
        type: integer
      createdAt:
        type: string
      errorMsg:
        type: string
      id:
        type: integer
      jsonResult:
        type: string
      method:
        type: string
      operIp:
        type: string
      operName:
        type: string
      operUserId:
        type: integer
      requestParam:
        type: string
      requestUrl:
        type: string
      status:
        type: integer
      title:
        type: string
    type: object
  github_com_fangzhengcn_workbackend_apps_backend_internal_vo.OperLogItem:
    properties:
      businessType:
        type: integer
      costTime:
        type: integer
      createdAt:
        type: string
      errorMsg:
        type: string
      id:
        type: integer
      method:
        type: string
      operIp:
        type: string
      operName:
        type: string
      operUserId:
        type: integer
      requestUrl:
        type: string
      status:
        type: integer
      title:
        type: string
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  github_com_fangzhengcn_workbackend_apps_backend_pkg_response.PageData:
    properties:
      list: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: 基于 RBAC 的通用后台管理系统接口
  title: 权限管理后台 API
  version: "1.0"
paths:
...
  /login-logs:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 日志ID集合
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.DeleteLogsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 删除登录日志
      tags:
      - 日志管理
    get:
      parameters:
      - in: query
        maxLength: 32
        name: beginTime
        type: string
      - in: query
        maxLength: 32
        name: endTime
        type: string
      - in: query
        maxLength: 64
        name: ipaddr
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 200
        minimum: 1
        name: size
        type: integer
      - enum:
        - 0
        - 1
        in: query
        name: status
        type: integer
      - in: query
        maxLength: 64
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_vo.LoginLogItem'
                        type: array
                    type: object
              type: object
      summary: 登录日志列表
      tags:
      - 日志管理
...
  /login-logs/export:
    get:
      parameters:
      - in: query
        maxLength: 32
        name: beginTime
        type: string
      - in: query
        maxLength: 32
        name: endTime
        type: string
      - in: query
        maxLength: 64
        name: ipaddr
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 200
        minimum: 1
        name: size
        type: integer
      - enum:
        - 0
        - 1
        in: query
        name: status
        type: integer
      - in: query
        maxLength: 64
        name: username
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: CSV 文件
          schema:
            type: file
      summary: 导出登录日志
      tags:
      - 日志管理
...
  /oper-logs:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 日志ID集合
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.DeleteLogsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 删除操作日志
      tags:
      - 日志管理
    get:
      parameters:
      - in: query
        maxLength: 32
        name: beginTime
        type: string
      - description: BusinessType 见 model.BusinessType* 常量：0 其他 1 新增 2 修改 3 删除 4 查询。
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        in: query
        name: businessType
        type: integer
      - in: query
        maxLength: 32
        name: endTime
        type: string
      - in: query
        maxLength: 64
        name: operName
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 200
        minimum: 1
        name: size
        type: integer
      - enum:
        - 0
        - 1
        in: query
        name: status
        type: integer
      - description: Title 是操作模块名（如「用户管理」），模糊匹配。
        in: query
        maxLength: 64
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_vo.OperLogItem'
                        type: array
                    type: object
              type: object
      summary: 操作日志列表
      tags:
      - 日志管理
  /oper-logs/{id}:
    get:
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_vo.OperLogDetail'
              type: object
      summary: 操作日志详情
      tags:
      - 日志管理
...
  /oper-logs/export:
    get:
      parameters:
      - in: query
        maxLength: 32
        name: beginTime
        type: string
      - description: BusinessType 见 model.BusinessType* 常量：0 其他 1 新增 2 修改 3 删除 4 查询。
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        in: query
        name: businessType
        type: integer
      - in: query
        maxLength: 32
        name: endTime
        type: string
      - in: query
        maxLength: 64
        name: operName
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 200
        minimum: 1
        name: size
        type: integer
      - enum:
        - 0
        - 1
        in: query
        name: status
        type: integer
      - description: Title 是操作模块名（如「用户管理」），模糊匹配。
        in: query
        maxLength: 64
        name: title
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: CSV 文件
          schema:
            type: file
      summary: 导出操作日志
      tags:
      - 日志管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"